Divide the serializer into admin and web app usage.
