+++ exams/api_admin/serializers.py
@@ -1,0 +1,245 @@
+from rest_framework import serializers
+
+from common.api.serializers import CreatorSerializer
+from enrollments.api.serializers import SessionSerializer
+from exams.api.serializers import ExamTemplateListSerializer
+from exams.models import Exam, ExamTemplate, ExamTemplateStatus, Question, Section
+
+from .validators import validate_gt_than_template_marks
+
+
+class QuestionCreateSerializer(serializers.ModelSerializer):
+    """Serializer when admin is creating a question."""
+
+    class Meta:
+        model = Question
+        fields = (
+            "id",
+            "detail",
+            "img",
+            "exam",
+            "section",
+            "feedback",
+        )
+
+    # def validate(self, attrs):
+
+    #     return
+
+
+class QuestionUpdateSerializer(serializers.ModelSerializer):
+    """Serializer when admin is updating a question."""
+
+    class Meta:
+        model = Question
+        fields = (
+            "id",
+            "detail",
+            "img",
+            "exam",
+            "section",
+            "feedback",
+        )
+        read_only_fields = (
+            "id",
+            "exam",
+            "section",
+        )
+
+    # def validate(self, attrs):
+
+    #     return
+
+
+def get_total_section_marks(template):
+    """Calculate total marks for the template.
+
+    Returns
+    -------
+    decimal
+        Total marks of the sections in the template.
+
+    """
+    return sum(section.get_section_marks() for section in template.sections.all())
+
+
+class SectionCRUDSerializer(serializers.ModelSerializer):
+    """CRUD serializer for Section."""
+
+    section_marks = serializers.SerializerMethodField(read_only=True)
+
+    class Meta:
+        model = Section
+        fields = (
+            "id",
+            "name",
+            "num_of_questions",
+            "pos_marks",
+            "neg_percentage",
+            "template",
+            "section_marks",
+        )
+
+    def get_section_marks(self, obj):
+        return obj.get_section_marks()
+
+    def validate(self, attrs):
+        template_instance = self.instance
+        template = attrs.get(
+            "template", template_instance and template_instance.template
+        )
+        pos_marks = attrs.get(
+            "pos_marks", template_instance and template_instance.pos_marks
+        )
+        num_of_questions = attrs.get(
+            "num_of_questions", template_instance and template_instance.num_of_questions
+        )
+
+        template_marks = template.full_marks
+        validate_gt_than_template_marks(template_marks, pos_marks)
+
+        section_marks = pos_marks * num_of_questions
+        validate_gt_than_template_marks(template_marks, section_marks)
+
+        return attrs
+
+
+class ExamTemplateCreateUpdateSerializer(CreatorSerializer):
+    """Exam Template Serializer."""
+
+    pass_marks = serializers.SerializerMethodField(read_only=True)
+
+    class Meta:
+        model = ExamTemplate
+        fields = CreatorSerializer.Meta.fields + (
+            "name",
+            "description",
+            "full_marks",
+            "pass_percentage",
+            "pass_marks",
+            "duration",
+            "display_num_questions",
+            "status",
+        )
+        read_only_fields = CreatorSerializer.Meta.read_only_fields
+
+    def update(self, instance, validated_data):
+        full_marks = validated_data.get("full_marks", instance.full_marks)
+        status = validated_data.get("status", instance.status)
+        total_section_marks = get_total_section_marks(instance)
+        print(f"full_marks: {full_marks}, total_section_marks: {total_section_marks}")
+        if (status == ExamTemplateStatus.COMPLETED) and (
+            full_marks != total_section_marks
+        ):
+            raise serializers.ValidationError(
+                (
+                    "Total marks should be equal to sum of section"
+                    + " marks on exam completion"
+                )
+            )
+        return super().update(instance, validated_data)
+
+    def get_pass_marks(self, obj):
+        return obj.pass_percentage * obj.full_marks
+
+
+class ExamTemplateRetrieveSerializer(CreatorSerializer):
+    """Serializer to retrieve exam template."""
+
+    pass_marks = serializers.SerializerMethodField(read_only=True)
+    sections = SectionCRUDSerializer(many=True, read_only=True)
+    residual_marks = serializers.SerializerMethodField(read_only=True)
+
+    class Meta:
+        model = ExamTemplate
+        fields = CreatorSerializer.Meta.fields + (
+            "name",
+            "description",
+            "full_marks",
+            "pass_percentage",
+            "pass_marks",
+            "duration",
+            "display_num_questions",
+            "sections",
+            "residual_marks",
+            "status",
+        )
+        read_only_fields = CreatorSerializer.Meta.read_only_fields
+
+    def get_pass_marks(self, obj):
+        return obj.pass_percentage * obj.full_marks
+
+    def get_residual_marks(self, obj):
+        sum_marks = get_total_section_marks(obj)
+        return obj.full_marks - sum_marks
+
+
+class ExamTemplateListAdminSerializer(ExamTemplateListSerializer):
+    class Meta:
+        model = ExamTemplate
+        fields = ExamTemplateListSerializer.Meta.fields
+
+
+class ExamCreateSerializer(CreatorSerializer):
+    """Serializer when admin is creating an exam."""
+
+    class Meta:
+        model = Exam
+        fields = CreatorSerializer.Meta.fields + (
+            "name",
+            "category",
+            "status",
+            "price",
+            "template",
+        )
+        read_only_fields = CreatorSerializer.Meta.read_only_fields + ("status",)
+
+
+class ExamListAdminSerializer(serializers.ModelSerializer):
+    """Serializer when user is listing exams."""
+
+    template = ExamTemplateListSerializer()
+
+    class Meta:
+        model = Exam
+        fields = (
+            "id",
+            "name",
+            "category",
+            "status",
+            "price",
+            "template",
+        )
+
+
+class ExamUpdateSerializer(CreatorSerializer):
+    """Serializer when admin is updating an exam."""
+
+    class Meta:
+        model = Exam
+        fields = CreatorSerializer.Meta.fields + (
+            "name",
+            "category",
+            "status",
+            "price",
+        )
+        read_only_fields = CreatorSerializer.Meta.read_only_fields + ("status",)
+
+
+class ExamRetrieveAdminSerializer(serializers.ModelSerializer):
+    """Serializer when admin is retrieving an exam."""
+
+    template = ExamTemplateListSerializer()
+    sessions = SessionSerializer(many=True, read_only=True)
+
+    class Meta:
+        model = Exam
+        fields = (
+            "id",
+            "name",
+            "category",
+            "status",
+            "price",
+            "template",
+            "sessions",
+        )
